# Avoid sycl waits in forwardEval code paths