--- a/src/neural/backends/sycl/kernels.h
+++ b/src/neural/backends/sycl/kernels.h
@@ -146,5 +146,9 @@ void inputPreprocessForAttentionBody(T* output, const T* input,
 template <typename T>
 void applyInputGating(T* output, const T* input, const T* mult, const T* add,
                       int N, int HW, int C, sycl::queue &sycl_queue);
+
+template <typename T>
+void genOffsetPointers(T** offsets, int heads, int max_batch, int depth,
+                       int d_model, T* k, T* q, T* b1, T* v, T* b2, sycl::queue &sycl_queue);
 }  // namespace sycldnn_backend
 }  // namespace lczero