@Menkib64 2025-08-03
I figures out my random test failures. The problem was me forgetting -ffast-math into compiler flags which breaks floating point handling of -inf.

@ContradNamiseb Would you test this, please?

Here is correct output from the optimised ./lc0 bench --threads=1 --task-workers=0  --nodes=1. Evaluation times keep improving because frequency scaling is that slow to react when GPU was idling.
Position: 1/10 rnbqkbnr/pppppppp/8/8/8/8/PPPPPPPP/RNBQKBNR w KQkq - 0 1
Benchmark time 123 ms, 1 nodes, -1 nps, move d2d4
bestmove d2d4

Position: 2/10 r3k2r/p1ppqpb1/bn2pnp1/3PN3/1p2P3/2N2Q1p/PPPBBPPP/R3K2R w KQkq - 0 10
Benchmark time 14 ms, 1 nodes, -1 nps, move e2a6
bestmove e2a6

Position: 3/10 8/2p5/3p4/KP5r/1R3p1k/8/4P1P1/8 w - - 0 11
Benchmark time 13 ms, 1 nodes, -1 nps, move b4f4
bestmove b4f4

Position: 4/10 4rrk1/pp1n3p/3q2pQ/2p1pb2/2PP4/2P3N1/P2B2PP/4RRK1 b - - 7 19
Benchmark time 12 ms, 1 nodes, -1 nps, move f5d3
bestmove f5d3

Position: 5/10 rq3rk1/ppp2ppp/1bnpb3/3N2B1/3NP3/7P/PPPQ1PP1/2KR3R w - - 7 14 moves d4e6
Benchmark time 11 ms, 1 nodes, -1 nps, move f7e6
bestmove f7e6

Position: 6/10 r1bq1r1k/1pp1n1pp/1p1p4/4p2Q/4Pp2/1BNP4/PPP2PPP/3R1RK1 w - - 2 14 moves g2g4
Benchmark time 10 ms, 1 nodes, -1 nps, move f4f3
bestmove f4f3

Position: 7/10 r3r1k1/2p2ppp/p1p1bn2/8/1q2P3/2NPQN2/PPP3PP/R4RK1 b - - 2 15
Benchmark time 10 ms, 1 nodes, -1 nps, move b4b2
bestmove b4b2

Position: 8/10 r1bbk1nr/pp3p1p/2n5/1N4p1/2Np1B2/8/PPP2PPP/2KR1B1R w kq - 0 13
Benchmark time 10 ms, 1 nodes, -1 nps, move f4g3
bestmove f4g3

Position: 9/10 r1bq1rk1/ppp1nppp/4n3/3p3Q/3P4/1BP1B3/PP1N2PP/R4RK1 w - - 1 16
Benchmark time 9 ms, 1 nodes, -1 nps, move a1e1
bestmove a1e1

Position: 10/10 4r1k1/r1q2ppp/ppp2n2/4P3/5Rb1/1N1BQ3/PPP3PP/R5K1 w - - 1 17
Benchmark time 9 ms, 1 nodes, -1 nps, move e3g3
bestmove e3g3