@Menkib64 2025-08-07
I decided to change the workgroup size calculation. I noticed that CPU generated buffer here is going to be problematic when trying to track dependencies for an out of order queue. Sycl is designed GPU->GPU operations easy to synchronize.